+++ accent/accent-common/src/main/java/de/embl/rieslab/accent/common/interfaces/Generator.java
@@ -4,7 +4,7 @@ import de.embl.rieslab.accent.common.data.calibration.Calibration;
 
 public interface Generator {
 	
-	public void generate(String path, Calibration calibration, Integer[] exposures);
+	public void generate(String path, Calibration calibration, double[] exposures);
 	
 	public boolean isRunning();
