+++ accent/accent-common/src/main/java/de/embl/rieslab/accent/common/interfaces/ui/GeneratePanelInterface.java
@@ -6,5 +6,5 @@ public interface GeneratePanelInterface {
 	
 	public void setProgress(String progress);
 	
-	public Integer[] getExposures();
+	public double[] getExposures();
 }
